backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.6.0
base58==2.1.1
blinker==1.5
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
cryptography==38.0.4
dbus-python==1.3.2
Deprecated==1.2.14
distro==1.8.0
distro-info==1.5+deb12u1
dnspython==2.6.1
ECPy==1.2.5
fastapi==0.112.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.4
httpx==0.24.1
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
keyring==23.9.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
markdown-it-py==2.1.0
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.2.2
pip==23.0.1
pycryptodome==3.20.0
pycurl==7.45.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pymongo==4.8.0
pynvim==0.4.2
pyparsing==3.0.9
PySimpleSOAP==1.16.2
python-apt==2.6.0
python-debian==0.1.49
python-debianbts==4.0.1
PyYAML==6.0
reportbug==12.0.0
requests==2.28.1
rich==13.3.1
SecretStorage==3.3.3
setuptools==66.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
types-Deprecated==1.2.9.20240311
typing_extensions==4.12.2
unattended-upgrades==0.1
urllib3==1.26.12
uvicorn==0.30.6
wadllib==1.3.6
websockets==11.0.3
wheel==0.38.4
wrapt==1.16.0
wsproto==1.2.0
xrpl-py==3.0.0
zipp==1.0.0